fix(lysis): widen ConstPool-index opcode fields to u32
The ConstPool is u32-indexed but LoadConst.idx, LoadInput.name_idx,
EmitAssertEqMsg.msg_idx and EmitWitnessCall{,Heap}.bytecode_const_idx
were u16 with a 2-byte wire encoding, and every walker emit site cast
the intern index `as u16`. Once the shared pool exceeds 65535 entries
the cast wraps; the wrapped index still passes the `idx < pool_len`
bounds check, so it resolves the wrong pool slot — a deep, scale-only
miscompile, or a validation failure when that slot is not a field
entry. Circuits with very large const pools (heavy bigint emulation)
hit this; smaller ones never reach the ceiling.

Widen those five fields to u32 (encode auto-widens via
u32::to_le_bytes; decode reads u32; redundant `as u32` casts dropped
in validate/execute). LoadCapture.idx stays u16 — it indexes the
runtime capture array, a different index space. ProgramBuilder
helpers and test/bench callers updated; the bad-visibility decode
fixture updated to the 4-byte name_idx layout. No bytecode VERSION
bump: lysis bytecode is encoded and decoded in-process in the same
call, never persisted or cross-version decoded.

Pinned by load_const_idx_survives_when_const_pool_exceeds_u16_ceiling
(asserts the emitted index exceeds the old u16 ceiling, resolves the
correct field, and round-trips through encode/decode/execute). The
ECDSAVerify boss-fight ignore reason is refreshed to present truth.

+++ lysis/benches/sha256_round.rs
@@ -41,7 +41,7 @@ fn build_sha256_round() -> lysis::Program<Bn254Fr> {
         .load_input(6, 7, Visibility::Public)
         .load_input(7, 8, Visibility::Public)
         .load_input(8, 9, Visibility::Witness)
-        .load_const(9, k_t as u16)
+        .load_const(9, k_t)
         .emit_add(10, 7, 9)
         .emit_add(11, 10, 8)
         .emit_add(12, 11, 4)
